view - add - edit - delete Product

+++ src/main/java/com/example/BuildPC/Service/ProductServiceImpl.java
@@ -4,12 +4,22 @@ import com.example.BuildPC.dtos.ProductDto;
 import com.example.BuildPC.models.Brand;
 import com.example.BuildPC.models.Category;
 import com.example.BuildPC.models.Product;
+import com.example.BuildPC.models.ProductImage;
 import com.example.BuildPC.repository.BrandRepository;
 import com.example.BuildPC.repository.CategoryRepository;
+import com.example.BuildPC.repository.ProductImageRepository;
 import com.example.BuildPC.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.FieldError;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.InputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,6 +33,8 @@ public class ProductServiceImpl implements ProductService{
     private CategoryRepository categoryRepository;
     @Autowired
     private BrandRepository brandRepository;
+    @Autowired
+    private ProductImageService productImageService;
 
     @Override
     public List<Product> findAll() {
@@ -43,6 +55,12 @@ public class ProductServiceImpl implements ProductService{
                 System.out.println("Brand not found");
                 return;
             }
+//            if(productRepository.existsByProductName(productDto.getProductName())){
+//                System.out.println("Product name already exists");
+//                return;
+//            }
+
+
             Product product = new Product();
             product.setProductName(productDto.getProductName());
             product.setProductSlug(productDto.getProductSlug());
@@ -52,9 +70,13 @@ public class ProductServiceImpl implements ProductService{
             product.setProductDesc(productDto.getProductDesc());
             product.setUnitsInStock(productDto.getUnitsInStock());
             product.setUnitsInOrder(productDto.getUnitsInOrder());
+            product.setProductStatus(productDto.isProductStatus());
             product.setBrand(brand.get());
 
-            productRepository.save(product);
+            Product saveProduct =  productRepository.save(product);
+
+            List<MultipartFile> images = productDto.getProductImages();
+            productImageService.createAllProductImages(images, saveProduct.getId());
         }catch (Exception e){
             System.out.println("Error while creating product" + e.getMessage());
         }
@@ -68,11 +90,30 @@ public class ProductServiceImpl implements ProductService{
 
     @Override
     public void updateProdcut(Product product) {
-        productRepository.save(product);
+        if(productRepository.existsById(product.getId())){
+            productRepository.save(product);
+        }else {
+            throw new RuntimeException("Product with id " + product.getId() + " does not exist");
+        }
+
     }
 
     @Override
     public void deleteProduct(int id) {
-        productRepository.deleteById(id);
+        Product product = productRepository.findById(id).orElse(null);
+        if(product != null) {
+            productImageService.deleteAllProductImages(product.getProductImages());
+            productRepository.deleteById(id);
+        }
     }
+
+        @Override
+        public boolean existsByProductName(String productName) {
+            return productRepository.existsByProductName(productName);
+        }
+
+//    @Override
+//    public List<Product> findByProductStatus() {
+//        return productRepository.findByProductStatus(true);
+//    }
 }
